// src/models/Skill/Skill_ID.ts
export enum SkillID {
    FLETCHING = 'Wood Working',
    WOODCUTTING = 'Wood Cutting',
    MINING = 'Mining',
    SMITHING = 'Smithing',
    FISHING = 'Fishing',
    CRAFTING = 'Crafting',
    GATHERING = 'Gathering',
    FIREMAKING = 'Firemaking',
    COOKING = 'Cooking',
    RUNECRAFTING = 'Runecrafting',
    DIVINATION = 'Divination',
    CONSTRUCTION = 'Construction',
    THIEVING = 'Thieving',
    POTION_MAKING = 'Potion Making',
    INVENTION = 'Invention',
    ENCHANTING = 'Enchanting',
    HUNTING = 'Hunting',
    WEAVING = 'Weaving',
    LEATHER_WORKING = 'Leather Working',
    AGILITY = 'Agility',
    HOLY = 'Holy',
    CURSE = 'Curse',
    FARMING = 'Farming',
    RANGED = 'Ranged'
}